HDDS-8215. Recon - SCM DB Insights.
