+++ hadoop-hdds/server-scm/src/main/java/org/apache/hadoop/hdds/scm/server/OzoneStorageContainerManager.java
@@ -18,16 +18,16 @@
 
 package org.apache.hadoop.hdds.scm.server;
 
+import java.io.IOException;
+import java.net.InetSocketAddress;
+
 import org.apache.hadoop.hdds.scm.block.BlockManager;
 import org.apache.hadoop.hdds.scm.container.ContainerManager;
-import org.apache.hadoop.hdds.scm.container.balancer.ContainerBalancer;
 import org.apache.hadoop.hdds.scm.container.replication.ReplicationManager;
+import org.apache.hadoop.hdds.scm.container.balancer.ContainerBalancer;
 import org.apache.hadoop.hdds.scm.ha.SCMNodeDetails;
 import org.apache.hadoop.hdds.scm.node.NodeManager;
 import org.apache.hadoop.hdds.scm.pipeline.PipelineManager;
-
-import java.io.IOException;
-import java.net.InetSocketAddress;
 
 /**
  * Interface for the SCM Facade class that can be used by a passive SCM like
@@ -58,5 +58,4 @@ public interface OzoneStorageContainerManager {
   InetSocketAddress getDatanodeRpcAddress();
 
   SCMNodeDetails getScmNodeDetails();
-
 }
